* This patch enables SLEEF to be referenced with GNU's vector ABI.
** SLEEF library with GNU ABI is completely separated from the library with the existing API. The reason for doing this is that if I only expose the GNU ABI version and somehow rename it in the header, SLEEF cannot be used with glibc, which already has the functions with the same names.
** In this patch, 1 ULP functions are default. Functions that return two values (like sincos) are not included in the GNU ABI version of SLEEF.
** The ABI specification does not provide a way to systematically specify accuracy via the function name. If we are going to allow the compiler automatically choose a function according to the compiler options, this might be a problem.
** We need a new testing program to confirm if this implementation is correct.
** If this patch is accepted, I am planning to add the library with SVML and LLVM naming conventions in the same way.

* This patch also enables make test to run on MSVC and Mac OSX.

* iutneon32 now automatically enables flushtozero mode for tester

* This patch fixes the following bugs.

** xsqrtf_u35 did not work correctly if compiled with VECEXT option
** iutsimd.c is now divided into two parts. This is needed to make it work with icc. In the previous implementation, the whole program is compiled with support for extended ISA (e.g. AVX512F), and this caused the IUT not to exit with the correct exit code when that ISA is not supported on the host.
** In tester, the reference ldexp function is now replaced with MPFR version.
** In tester2simddp.c, function vgeti was using syntax incompatible with MSVC.
** iutsimd did not work correctly if compiled with icc due to incorrect length of array in func_d_d_i macro.
** AVX512F is now enabled if GCC version is 5 or later
