+++ lib/Makefile.vc
@@ -1,17 +1,25 @@
 .PHONY: all
-all : libsleef libsleefdft
+all : libsleef libsleefgnuabi libsleefdft
 
 .PHONY: libsleef
 libsleef : libsleef.dll
 
+.PHONY: libsleefgnuabi
+libsleefgnuabi : libsleefgnuabi.dll
+
 .PHONY: libsleef-dft
 libsleefdft : libsleefdft.dll
 
-libsleef.dll : ../src/libm/libm.lib ../src/common/common.lib
-	link /DLL /OUT:libsleef.dll ../src/libm/*.obj ../src/common/common.obj
+libsleef.dll : ../src/libm/libm.lib ../src/libm/OBJ.txt ../src/common/common.lib
+	link /DLL /OUT:libsleef.dll `cat ../src/libm/OBJ.txt` ../src/common/common.obj
+
+libsleefgnuabi.dll : ../src/libm/libgm.lib ../src/libm/OBJ2.txt
+	link /DLL /OUT:libsleefgnuabi.dll `cat ../src/libm/OBJ2.txt`
 
 libsleefdft.dll : ../src/dft/dft.lib libsleef.dll ../src/common/arraymap.lib
 	link /DLL /OUT:libsleefdft.dll ../src/dft/*.obj ../src/common/arraymap.obj libsleef.lib
+
+sleef.h : ../src/libm/sleef.h
 
 #libsleefstatic.lib : ../src/libm/libm.lib ../src/common/common.lib ../src/libm/sleef.h
 #	lib /out:libsleefstatic.lib ../src/libm/*.obj ../src/common/common.obj
@@ -21,6 +29,9 @@ libsleefdft.dll : ../src/dft/dft.lib libsleef.dll ../src/common/arraymap.lib
 
 ../src/libm/libm.lib :
 	+"$(MAKE)" --directory=../src/libm -f Makefile.vc libm.lib
+
+../src/libm/libgm.lib :
+	+"$(MAKE)" --directory=../src/libm -f Makefile.vc libgm.lib
 
 ../src/libm/sleef.h :
 	+"$(MAKE)" --directory=../src/libm -f Makefile.vc sleef.h
@@ -33,6 +44,12 @@ libsleefdft.dll : ../src/dft/dft.lib libsleef.dll ../src/common/arraymap.lib
 
 ../src/common/arraymap.lib :
 	+"$(MAKE)" --directory=../src/common -f Makefile.vc arraymap.lib
+
+../src/libm/OBJ.txt :
+	+"$(MAKE)" --directory=../src/libm -f Makefile.vc OBJ.txt
+
+../src/libm/OBJ2.txt :
+	+"$(MAKE)" --directory=../src/libm -f Makefile.vc OBJ2.txt
 
 .PHONY: clean
 clean :
