+++ src/arch/helperavx2.h
@@ -43,7 +43,9 @@ typedef __m256i vint2;
 
 //
 
+#ifndef __SLEEF_H__
 void Sleef_x86CpuID(int32_t out[4], uint32_t eax, uint32_t ecx);
+#endif
 
 static int cpuSupportsAVX2() {
     int32_t reg[4];
