+++ src/libm-tester/Makefile.vc
@@ -3,42 +3,54 @@ CFLAGS+=/fp:strict /I..\common /I..\arch /I../../include /I../libm
 
 #
 
-TARGET = iut.exe iutsse2.exe iutavx.exe iutavx2.exe
+TARGET_IUT = iut.exe iutsse2.exe iutavx.exe iutavx2.exe
 
 .PHONY: all
-all : $(TARGET)
+all : $(TARGET_IUT)
+
+test : $(TARGET_IUT) tester
+	cp ../../lib/*.dll .
+	$(foreach var,$(TARGET_IUT), ./tester ./$(var) || exit;)
+
+tester : tester.c testerutil.c
+	+"$(MAKE)" --directory=. tester
 
 #
 
-iut.exe : ../../lib/libsleef.lib ../../include/sleef.h iut.c ../libm/rename.h ../../include/sleef.h
-	$(CC) $(CFLAGS) iut.c ../../lib/libsleef.lib
+iutsimdmain.obj : iutsimdmain.c
+	$(CC) $(CFLAGS) iutsimdmain.c /c /Fo iutsimdmain.obj
 
-iutsse2.exe : ../../lib/libsleef.lib ../../include/sleef.h iutsimd.c ../libm/renamesse2.h ../../include/sleef.h
-	$(CC) $(CFLAGS) /DENABLE_SSE2 /D__SSE2__ /DLOGVECTLENDP=1 /DLOGVECTLENSP=2 iutsimd.c ../../lib/libsleef.lib /Feiutsse2.exe
+#
+
+iut.exe : ../../lib/libsleef.lib ../../include/sleef.h iut.c testerutil.c ../libm/rename.h ../../include/sleef.h
+	$(CC) $(CFLAGS) iut.c testerutil.c ../../lib/libsleef.lib
+
+iutsse2.exe : ../../lib/libsleef.lib ../../include/sleef.h iutsimdmain.obj iutsimd.c testerutil.c ../libm/renamesse2.h ../../include/sleef.h
+	$(CC) $(CFLAGS) /DENABLE_SSE2 /D__SSE2__ /DLOGVECTLENDP=1 /DLOGVECTLENSP=2 iutsimdmain.obj iutsimd.c testerutil.c ../../lib/libsleef.lib /Feiutsse2.exe
 
 ../libm/renamesse2.h :
 	+"$(MAKE)" --directory=../libm -f Makefile.vc renamesse2.h
 
 #
 
-iutavx.exe : ../../lib/libsleef.lib ../../include/sleef.h iutsimd.c ../libm/renameavx.h ../../include/sleef.h
-	$(CC) $(CFLAGS) /DENABLE_AVX /DLOGVECTLENDP=2 /DLOGVECTLENSP=4 /arch:AVX iutsimd.c ../../lib/libsleef.lib /Feiutavx.exe
+iutavx.exe : ../../lib/libsleef.lib ../../include/sleef.h iutsimdmain.obj iutsimd.c testerutil.c ../libm/renameavx.h ../../include/sleef.h
+	$(CC) $(CFLAGS) /DENABLE_AVX /DLOGVECTLENDP=2 /DLOGVECTLENSP=4 /arch:AVX iutsimdmain.obj iutsimd.c testerutil.c ../../lib/libsleef.lib /Feiutavx.exe
 
 ../libm/renameavx.h :
 	+"$(MAKE)" --directory=../libm -f Makefile.vc renameavx.h
 
 #
 
-iutavx2.exe : ../../lib/libsleef.lib ../../include/sleef.h iutsimd.c ../libm/renameavx2.h ../../include/sleef.h
-	$(CC) $(CFLAGS) /DENABLE_AVX2 /DLOGVECTLENDP=2 /DLOGVECTLENSP=4 /arch:AVX2 iutsimd.c ../../lib/libsleef.lib /Feiutavx2.exe
+iutavx2.exe : ../../lib/libsleef.lib ../../include/sleef.h iutsimdmain.obj iutsimd.c testerutil.c ../libm/renameavx2.h ../../include/sleef.h
+	$(CC) $(CFLAGS) /DENABLE_AVX2 /DLOGVECTLENDP=2 /DLOGVECTLENSP=4 /arch:AVX2 iutsimdmain.obj iutsimd.c testerutil.c ../../lib/libsleef.lib /Feiutavx2.exe
 
 ../libm/renameavx2.h :
 	+"$(MAKE)" --directory=../libm -f Makefile.vc renameavx2.h
 
 #
 
-iutpurec.exe : ../libm/sleefsimddp.c ../libm/sleefsimdsp.c iutsimd.c
-	$(CC) $(CFLAGS) /DENABLE_PUREC ../libm/sleefsimddp.c ../libm/sleefsimdsp.c iutsimd.c /Feiutpurec.exe
+iutpurec.exe : ../libm/sleefsimddp.c ../libm/sleefsimdsp.c iutsimdmain.obj iutsimd.c testerutil.c
+	$(CC) $(CFLAGS) /DENABLE_PUREC ../libm/sleefsimddp.c ../libm/sleefsimdsp.c iutsimdmain.obj iutsimd.c testerutil.c /Feiutpurec.exe
 
 #
