--- src/libm-tester/starttester2.sh
@@ -1,3 +1,0 @@
-#!/bin/sh
-nohup nice ./tester2 >> ../tester2.out &
-nohup nice ./tester2f >> ../tester2.out &
