--- src/libm-tester/stoptester2.sh
@@ -1,2 +1,0 @@
-#!/bin/sh
-killall tester2 tester2f
