+++ src/libm-tester/tester2simdsp.c
@@ -22,10 +22,6 @@
 #include "sleef.h"
 #include "testerutil.h"
 
-#if defined(__APPLE__)
-static int isinff(float x) { return x == __builtin_inff() || x == -__builtin_inff(); }
-#endif
-
 #define DORENAME
 
 #ifdef ENABLE_SSE2
@@ -890,7 +886,7 @@ int main(int argc,char **argv)
       double u0 = countULP2sp(t = vget(xfmaf(vd, vd2, vd3), e), frx);
       double c = mpfr_get_d(frx, GMP_RNDN);
 
-      if ((-1e+34 < c && c < 1e+34 && u0 > 0.5001) ||
+      if ((-1e+34 < c && c < 1e+33 && u0 > 0.5001) ||
 	  !(u0 <= 0.5001 || isinf(t))) {
 	printf(ISANAME " fmaf arg=%.20g, %.20g, %.20g  ulp=%.20g\n", d, d2, d3, u0);
 	printf("correct = %.20g, test = %.20g\n", mpfr_get_d(frx, GMP_RNDN), t);
