+++ src/libm-tester/testerutil.c
@@ -20,6 +20,8 @@
 #include <signal.h>
 #endif
 
+#include "misc.h"
+
 #define DENORMAL_DBL_MIN (4.9406564584124654418e-324)
 #define POSITIVE_INFINITY INFINITY
 #define NEGATIVE_INFINITY (-INFINITY)
